Fixed logic that ensures that status bar is not hidden.
Fixed code that makes sure that the notification does not hide the status bar.
The original accounted for the status bar height in
+[AJNotificationView showNoticeInView:type:title:linedBackground:hideAfter:offset:]
but didn't actually user the status bar height that was calculated.

Also fixed a logic error in the calculation of status bar height. The original
failed when the view passed was not a UIWindow. The original accounted for
the status bar height whenever the view passed was not a UIWindow.

Removed duplicate code from
+[AJNotificationView showNoticeInView:type:type title:title linedBackground:hideAfter:].