Added tappable notification and a delay to notifications.
Added the ability to define a response block that is called when a notification
banner is tapped by the user.

Added the ability to delay the display of a notification banner. This might
be useful when a notification is displayed when the application becomes
active. Without the delay, the banner is already displayed when the application
launches. Delaying the notification by a fraction of a second allows the
notification animation to be viewed by the user so that it can catch his
attention.